moved stories to open / manage folders per variant
